created mapsReducer module and corrispetive test file
