+++ src/redux/reducers/rootReducer.js
@@ -2,11 +2,12 @@ import { combineReducers } from 'redux';
 import * as reducers from './reducers';
 import { locationsReducer } from './locationsReducer'
 import { singleLocationReducer } from './singleLocationReducer'
+import { mapsReducer } from './mapsReducer'
 
 const rootReducer = combineReducers({
   locations: locationsReducer,
   newLocation: reducers.addLocationReducer,
-  maps: reducers.mapsReducer,
+  maps: mapsReducer,
   location: singleLocationReducer,
   activeLocation: reducers.activeLocation,
   auth: reducers.authReducer,
